Move SubmenuDeepFind to module and fix ShowAppMenu
