+++ src/js/InAppList.js
@@ -3,36 +3,7 @@ import React from 'react';
 import AppHeader from './containers/Header';
 import { VSubMenu, HSubMenu } from './containers/SubMenu';
 import store from './store'
-
-// Depth First Search Recursive Function
-function SubmenuDeepFind(menu, parentID, depth) { 
-    if (!menu || !parentID) {
-        return null;
-    }
-    var deepSubMenu = null;
-    var subMenu = menu.find((command) => {
-        if (command.subMenu) { 
-            var result = SubmenuDeepFind(command.subMenu, parentID, depth++)
-            if (result && result.subMenu) {
-                deepSubMenu = result;
-                return true;
-            }
-        }
-        return command.menuID === parentID
-    });
-
-    if (deepSubMenu) {
-        return deepSubMenu;
-    }
-    
-    if (subMenu) {
-        return {
-            subMenu: subMenu,
-            depth: depth
-        }
-    }
-    return null;
-}
+import SubmenuDeepFind from './Utils/SubMenuDeepFind'
 
 export default class InAppMenu extends React.Component {
     render() {
