+++ src/js/containers/Header.js
@@ -1,32 +1,6 @@
 import { connect } from 'react-redux'
 import AppHeader from '../AppHeader'
-
-function SubmenuDeepFind(menu, parentID, depth) { 
-    if (!menu || !parentID) {
-        return null;
-    }
-    var deepSubMenu = null;
-    var subMenu = menu.find((command) => {
-        if (command.subMenu) { 
-            var result = SubmenuDeepFind(command.subMenu, parentID, depth++)
-            if (result && result.subMenu) {
-                deepSubMenu = result;
-                return true;
-            }
-        }
-        return command.menuID === parentID
-    });
-    if (deepSubMenu) {
-        return deepSubMenu;
-    }
-    if (subMenu) {
-        return {
-            subMenu: subMenu,
-            depth: depth
-        }
-    }
-    return null;
-}
+import SubmenuDeepFind from '../Utils/SubMenuDeepFind'
 
 const mapStateToProps = (state) => {
     var activeApp = state.activeApp
